requirements.txt
beautifulsoup4==4.14.3
celery>=5.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
dnspython==2.8.0
fastapi>=0.109.0
idna==3.11
pymongo==4.16.0
python-dotenv==1.2.1
redis>=5.0.0
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn>=0.27.0
